refactor(naming): normalize manifest and lockfile modules
