--- src/lib/rapm_lock/mod.rs
@@ -1,2 +1,0 @@
-mod constraint;
-pub mod lockfile;
